birch.df {ggpp}	R Documentation

Birch seedlings' size

Description

A dataset containing the measurements on 350 birch seedlings.

Usage

birch.df

birch_dw.df

Format

A data.frame object with 350 rows and 8 variables.

A data.frame object with 700 rows and 5 variables.

Details

The data are for seedlings grown in trays with cells or containers of two different volumes. For each of these types of trays, all cells, 1/2 of the cells or 1/4 of the cells contained seedlings. Root-collar diameter (mm), height (cm), dry mass (mg) of stems and roots. Measurements done at the end of the first growing season, after leaf fall.

References

Aphalo, P. J. and Rikala, R. (2003) Field performance of silver-birch planting-stock grown at different spacing and in containers of different volume. New Forests, 25:93-108. doi: 10.1023/A:1022618810937.

See Also

Other Plant growth and morphology data: ivy.df

Examples

colnames(birch.df)
head(birch.df)

colnames(birch_dw.df)
head(birch_dw.df)

[Package ggpp version 0.6.0 Index]