check_default_colour {ggpp}	R Documentation

Check default.colour argument

Description

if possible use ink from theme as default colour

Usage

check_default_colour(default.colour, theme = NULL)

Arguments

default.colour
The value to be checked

theme
The theme from where to fetch a replacement value when default.colour is NULL.

Details

The replacement value is fetched from the theme only if 'ggplot2' >= 4.0.0 and otherwise "black" is used. NA values are passed through as they can be used to disable rendering of grob elements.

Value

A colour definition or NA

[Package ggpp version 0.6.0 Index]