ivy.df {ggpp}	R Documentation

Ivy photosynthesis light response

Description

A dataset containing photosynthesis measurements on four ivy plants.

Usage

ivy.df

Format

A data.frame object with 36 rows and 6 variables.

Details

For each plant a light response curve of photosynthesis was measured using a custom-built system and software that allowed controlling the concentrations of water vapour and carbon dioxide at the surface of the leaves, i.e., inside the air boundary layer.

References

Aphalo, P. J. (1991) Interactions in Stomatal Function. PhD thesis, University of Edinburgh. http://hdl.handle.net/1842/14758.

See Also

Other Plant growth and morphology data: birch.df

Examples

colnames(ivy.df)
head(ivy.df)

[Package ggpp version 0.6.0 Index]