scale_continuous_npc {ggpp}	R Documentation

Position scales for continuous data (npcx & npcy)

Description

scale_npcx_continuous() and scale_npcy_continuous() are scales for continuous npcx and npcy aesthetics expressed in "npc" units. There are no variants. Obviously limits are always the full range of "npc" units and transformations meaningless. These scales are used by the newly defined aesthetics npcx and npcy.

Usage

scale_npcx_continuous(...)

scale_npcy_continuous(...)

Arguments

...
Other arguments passed on to continuous_scale()

Value

A "Scale" object.

[Package ggpp version 0.6.0 Index]