these2logical {ggpp}	R Documentation

Convert keep.these argument into logical vector

Description

Convert keep.these argument into logical vector

Usage

these2logical(these, data, these.target = "label")

Arguments

these
character vector, integer vector, logical vector or function that takes the variable mapped to the label aesthetic as first argument and returns a character vector or a logical vector. These rows from data are selected irrespective of the local density.

data
data.frame The plot layer's data set.

these.target
character The name of the variable (or aesthetic) in which to select the target passed through these.

[Package ggpp version 0.6.0 Index]