volcano_example.df {ggpp}	R Documentation

Gene expression data

Description

A dataset containing reshaped and simplified output from an analysis of data from RNAseq done with package edgeR. Original data from gene expression in the plant species Arabidopsis thaliana.

Usage

volcano_example.df

Format

A data.frame object with 1218 rows and 5 variables

References

Rai, Neha; O'Hara, Andrew; Farkas, Daniel; Safronov, Omid; Ratanasopa, Khuanpiroon; Wang, Fang; Lindfors, Anders V.; Jenkins, Gareth I.; Lehto, Tarja; Salojärvi, Jarkko; Brosché, Mikael; Strid. Åke; Aphalo, Pedro José; Morales, Luis Orlando (2020) The photoreceptor UVR8 mediates the perception of both UV-B and UV-A wavelengths up to 350 nm of sunlight with responsivity moderated by cryptochromes. Plant, Cell & Environment, 43:1513-1527.

See Also

Other Transcriptomics data: quadrant_example.df

Examples

colnames(volcano_example.df)
head(volcano_example.df)

[Package ggpp version 0.6.0 Index]